requirements.txt
requests
pytest
python-dotenv==1.0.0
sqlmodel
